Done Account

+++ sourceFE/src/pages/dashboard/AdminDashboard.jsx
@@ -1,4 +1,5 @@
-import React, { useEffect, useState } from "react";
+import React, { useEffect, useRef, useState } from "react";
+import { Link } from "react-router-dom";
 import { deleteAccount, getAccounts, updateAccount } from "../../ultis/utilsAccount";
 import useAccount from "../../hooks/useAccount";
 import { list } from "postcss";
@@ -13,17 +14,28 @@ import {
   SearchIcon,
 } from "../../components/headers/icon.jsx";
 import Modal from "react-modal";
+
+
 const AdminDashboard = () => {
   const { listAccount, setListAccount, page, setPage, filter, setFilter } =
     useAccount();
-
+  // const handleFilter = (list) =>{
+  //   return list.filter((value)=>(value?.role?.name !== 'admin' ));
+     
+  // }
   const [isLoading, setIsLoading] = useState(true);
   useEffect(() => {
     getAccounts(filter)
       .then((res) => {
         console.log("call api", res);
+        // const listAcc  = handleFilter(res.accounts)
+        // console.log('list',listAcc);
         setListAccount(res.accounts);
         setIsLoading(false);
+        setPage({
+          totalPages :res.page.totalPages,
+          currentPage :res.page.currentPage
+        })
       })
       .catch((err) => {
         alert(
@@ -31,10 +43,10 @@ const AdminDashboard = () => {
         );
         setIsLoading(false);
       });
-  }, []);
-
+  }, [filter]);
 
   
+  const searchInputRef = useRef(null);
   const [isShowModal, setIsShowModal] = useState(false);
   const [userSelect, setUserSelect] = useState({});
   function openModal(userInfor) {
@@ -62,7 +74,7 @@ const AdminDashboard = () => {
     };
     updateAccount(acc)
       .then((res) => {
-        console.log(res.data);
+        // console.log(res.data);
         if(res.status == 200)
         {
           setListAccount((prev) =>
@@ -78,19 +90,34 @@ const AdminDashboard = () => {
       });
   };
   const handleDeleteAcc = (acc) =>{
-    deleteAccount(acc._id).then((res)=>{
-      if(res.status == 200 || res.status == 202)
-      {
-        setListAccount((prev)=> prev.filter((account)=>
-          account._id !== acc._id
-        ))
+    const userConfirmed = window.confirm("Are you sure you want to delete this account?");
+
+  // If the user confirmed, proceed with the deletion
+  if (userConfirmed) {
+    deleteAccount(acc._id).then((res) => {
+      if (res.status === 200 || res.status === 202) {
+        setListAccount((prev) => prev.filter((account) => account._id !== acc._id));
       }
-        
-    }).catch((err)=>{
-      alert('Could not remove this account')
-    })
+    }).catch((err) => {
+      alert('Could not remove this account');
+    });
   }
-  console.log("userselect", userSelect);
+  }
+  const handlePageChange = (newPage) => {
+    setFilter((prevFilter) => ({
+      ...prevFilter,
+      page: newPage,
+    }));
+  };
+  // const handleSearch = () => {
+  //   const query = searchInputRef.current.value;
+  //   console.log('listacc',listAccount);
+  //   const filtered = listAccount.filter(account => 
+  //     account?.name.toLowerCase().includes(query.toLowerCase())
+  //   );
+  //   setListAccount(filtered);
+  // };
+  // console.log("userselect", userSelect);
   const customStyles = {
     content: {
       top: "55%",
@@ -102,35 +129,36 @@ const AdminDashboard = () => {
     },
   };
 
-  console.log(listAccount);
+  // console.log(listAccount);
   return !isLoading ? (
-    <div className=" flex flex-col justify-start items-center mx-auto bg-slate-200  w-screen  max-w-[1080px] ">
+    <div className=" flex flex-col justify-start items-center mx-auto bg-slate-200  w-screen  max-w-[1080px]  ">
       <div className="w-[1020px] mt-2 h-[60px] flex items-center justify-between  bg-slate-500 dark:bg-gray-700 p-4 rounded-t-lg shadow">
         <div className="flex-none w-[20px] h-[20px]"></div>
         <div className="shrink w-[300px] h-[40px]">
           <div className="flex bg-gray-200 rounded-lg px-4 py-2">
-            <div className="mx-2">
+            <button className="mx-2" >
               <SearchIcon classname="text-gray-500 hover:bg-gray-400" />
-            </div>
+            </button>
             <input
               type="text"
               placeholder="Search"
+              ref={searchInputRef} 
               className="bg-transparent border-none outline-none"
             />
             {/* <i className="fas fa-search text-gray-500"></i> */}
           </div>
         </div>
         <div className="flex-none w-[148px] h-[40px] ">
-          <button className="flex items-center bg-blue-500 text-white px-4 py-2 rounded-lg shadow hover:bg-blue-600">
+          <Link to={"/addAccount"} className="flex items-center bg-blue-500 text-white px-4 py-2 rounded-lg shadow hover:bg-blue-600">
             {/* <i className="fas fa-plus-circle mr-2"></i> */}
             <div className="mx-1">
               <AddUserIcon />
             </div>
             Add Customer
-          </button>
+          </Link>
         </div>
       </div>
-      <div className="bg-while-300 w-[1020px] ">
+      <div className="bg-while-300 w-[1020px] min-h-screen ">
         <div className="relative overflow-x-auto shadow-md sm:rounded-b-lg ">
           <table className="w-full text-sm text-left rtl:text-right text-gray-500 dark:text-gray-400 ">
             <thead className="h-[50px] text-xs text-gray-700 uppercase bg-slate-200 dark:bg-gray-700 dark:text-gray-400">
@@ -156,7 +184,7 @@ const AdminDashboard = () => {
               </tr>
             </thead>
             <tbody>
-              {listAccount?.map((account, index) => (
+              {listAccount && listAccount?.map((account, index) => (
                 <tr
                   key={account._id}
                   className="odd:bg-white odd:dark:bg-gray-900 even:bg-gray-50 even:dark:bg-gray-800 border-b dark:border-gray-700"
@@ -165,7 +193,7 @@ const AdminDashboard = () => {
                     scope="row"
                     className="px-6 py-4 font-medium text-gray-900 whitespace-nowrap dark:text-white"
                   >
-                    {account?.username}
+                    {account?.username}<p className="text-blue-600 text-xs">{account?.role?.name}</p>
                   </th>
                   <td className="px-6 py-4">{account?.email}</td>
                   <td className="px-6 py-4">{account?.name}</td>
@@ -176,11 +204,16 @@ const AdminDashboard = () => {
                     {formatDate(account?.updatedAt)}
                   </td>
                   <td className="px-6 py-4 flex flex-row justify-center item-center">
-                    <div
+                    {account?.role?.name==='admin' ? (
+                    <div>
+
+                    </div>):(
+                      <>
+                      <div
                       onClick={() => {
                         openModal(account);
                       }}
-                      className="mt-4 font-medium text-blue-300 dark:text-blue-500 hover:text-blue-700 "
+                      className="mt-2 font-medium text-blue-300 dark:text-blue-500 hover:text-blue-700 "
                     >
                       <EditIcon></EditIcon>
                     </div>
@@ -188,10 +221,12 @@ const AdminDashboard = () => {
                       onClick={()=>{
                         handleDeleteAcc(account)
                       }}
-                      className="mx-2 mt-4 font-medium text-red-300 dark:text-red-500 hover:text-red-700"
+                      className="mx-2 mt-2 font-medium text-red-300 dark:text-red-500 hover:text-red-700"
                     >
                       <DeleteIcon></DeleteIcon>
                     </button>
+                      </>
+                    )}
                   </td>
                 </tr>
               ))}
@@ -201,20 +236,19 @@ const AdminDashboard = () => {
                 <td colSpan="6" className="px-6 py-3">
                   <div className="flex justify-end  items-center">
                     <button
-                      // disabled={currentPage === 1}
-                      // onClick={() => handlePageChange(currentPage - 1)}
-                      className="px-4 py-2 bg-gray-200 text-gray-700 rounded disabled:opacity-50 hover:bg-green-300 mx-2"
+                      disabled={page?.currentPage === 1}
+                      onClick={() => handlePageChange(page?.currentPage - 1)}
+                      className="px-4 py-2 bg-gray-200 text-gray-700 rounded disabled:opacity-50  mx-2"
                     >
                       <LeftIcon />
                     </button>
                     <span>
-                      {/* Page {currentPage} of {totalPages} */}
-                      Page 1 of 10
+                      Page {page?.currentPage} of {page?.totalPages}
                     </span>
                     <button
-                      // disabled={currentPage === totalPages}
-                      // onClick={() => handlePageChange(currentPage + 1)}
-                      className="px-4 py-2 bg-gray-200 text-gray-700 rounded disabled:opacity-50 hover:bg-green-300 mx-2"
+                      disabled={page?.currentPage === page?.totalPages}
+                      onClick={() => handlePageChange(page?.currentPage + 1)}
+                      className="px-4 py-2 bg-gray-200 text-gray-700 rounded disabled:opacity-50  mx-2"
                     >
                       <RightIcon />
                     </button>
